requirements.txt
streamlit==1.47.0
pillow==11.3.0
numpy==2.0.2
opencv-python==4.10.0
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.166
deep-sort-realtime==1.3.2
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.6.1
pyyaml==6.0.2
scipy==1.13.1